server/requirements.txt
fastapi[standard]
uvicorn
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-multipart
motor
pydantic[email]
pydantic-settings
celery
redis
kombu
requests
langchain
langchain-groq
gunicorn